+++ bin/front-linter.js
@@ -3,6 +3,7 @@
 const program = require('commander');
 
 const pkg = require('../package.json');
+
 const version = pkg.version;
 
 program.version(version, '    --version');